Moving files from root to internal/game
