--- campaign.go
@@ -1,4 +1,0 @@
-package main
-
-type campaign struct {
-}
