--- equipment/slot/slot.go
@@ -1,23 +1,0 @@
-package slot
-
-// Location represents a specific equipmet slot.
-type Location int
-
-const (
-	Arms Location = iota
-	Back
-	Body
-	Face
-	Feet
-	Finger1
-	Finger2
-	Hand
-	Head
-	Legs
-	Necklace
-	Shield
-	Shoulders
-	Waist
-	Weapon
-	Wrist
-)
