--- equipment/weapon.go
@@ -1,71 +1,0 @@
-package equipment
-
-var (
-	offenseAttrib = []attribute{
-		{"Polished", 1},
-		{"Serrated", 1},
-		{"Heavy", 1},
-		{"Pronged", 2},
-		{"Steely", 2},
-		{"Vicious", 3},
-		{"Venomed", 4},
-		{"Stabbity", 4},
-		{"Dancing", 5},
-		{"Invisible", 6},
-		{"Vorpal", 7},
-	}
-
-	offenseAttribBad = []attribute{
-		{"Dull", -2},
-		{"Tarnished", -1},
-		{"Rusty", -3},
-		{"Padded", -5},
-		{"Bent", -4},
-		{"Mini", -4},
-		{"Rubber", -6},
-		{"Nerf", -7},
-		{"Unbalanced", -2},
-	}
-
-	weapons = []equipment{
-		{"Stick", 0},
-		{"Broken Bottle", 1},
-		{"Shiv", 1},
-		{"Sprig", 1},
-		{"Oxgoad", 1},
-		{"Eelspear", 2},
-		{"Bowie Knife", 2},
-		{"Claw Hammer", 2},
-		{"Handpeen", 2},
-		{"Andiron", 3},
-		{"Hatchet", 3},
-		{"Tomahawk", 3},
-		{"Hackbarm", 3},
-		{"Crowbar", 4},
-		{"Mace", 4},
-		{"Battleadze", 4},
-		{"Leafmace", 5},
-		{"Shortsword", 5},
-		{"Longiron", 5},
-		{"Poachard", 5},
-		{"Baselard", 5},
-		{"Whinyard", 6},
-		{"Blunderbuss", 6},
-		{"Longsword", 6},
-		{"Crankbow", 6},
-		{"Blibo", 7},
-		{"Broadsword", 7},
-		{"Kreen", 7},
-		{"Warhammer", 7},
-		{"Morning Star", 8},
-		{"Pole-adze", 8},
-		{"Spontoon", 8},
-		{"Bastard Sword", 9},
-		{"Peen-arm", 9},
-		{"Culverin", 10},
-		{"Lance", 10},
-		{"Halberd", 11},
-		{"Poleax", 12},
-		{"Bandyclef", 15},
-	}
-)
