+++ internal/game/campaign.go
@@ -1,0 +1,4 @@
+package game
+
+type campaign struct {
+}
