+++ internal/game/equipment/equipment_test.go
