+++ internal/game/player_test.go
@@ -1,4 +1,4 @@
-package main
+package game
 
 import "testing"
