--- spells.go
@@ -1,10 +1,0 @@
-package main
-
-// Spells is a slice of Spell
-type Spells []Spell
-
-// Spell represents the information for a spell in the players spellbook.
-type Spell struct {
-	Name  string
-	Level uint64
-}
